Add the ability to specify a host for the loader
It could be useful to have the ability to look for config files
on a remote host. This gives you the possibility to configure a
custom host class, which performs actions on a remote target.

Todo: loader.dump should create the tempfile on the remote host.
Document the new capabilities.