Heredocs should preserve whitespace with line breaks. (#84)
Don't normalize value for heredocs

The perl parser normalizes spaces between words when they're split on
several lines. We shouldn't do that for heredocs -- the logic for this
is moved down to the lexer since the parser doesn't know whether
the underlying value was defined by a heredoc.

This commit also adds tests for heredocs preserving whitespace.